requirements.txt
Flask>=2.0.0
pandas>=1.3.0
numpy>=1.21.0
scipy>=1.7.0
requests>=2.26.0
python-dotenv>=0.19.0
websocket-client>=1.0.0
